+++ dl/transform/image.py
@@ -8,8 +8,9 @@ class RandomTransformation:
     def __init__(self, magnitude=1, seed=1001):
         self._magnitude = magnitude
         self._rng = np.random.default_rng(seed)
-    
+            
     def _get_random_values(self, x):
+        # We increase the translation for bigger images
         translation_range = 4 * min(x.shape[-2:]) // 256
         translation = self._rng.uniform(-translation_range * self._magnitude, 
                                         translation_range * self._magnitude, (1, 2))
@@ -18,9 +19,9 @@ class RandomTransformation:
         angle = self._rng.uniform(-2 * self._magnitude, 
                                   2 * self._magnitude, (1,))
 
-        translation = tensor_utils.tensor(translation, device=x.device)
-        scale = tensor_utils.tensor(scale, device=x.device)
-        angle = tensor_utils.tensor(angle, device=x.device)
+        translation = tensor_utils.tensor(translation, dtype=x.dtype, device=x.device)
+        scale = tensor_utils.tensor(scale, dtype=x.dtype, device=x.device)
+        angle = tensor_utils.tensor(angle, dtype=x.dtype, device=x.device)
         
         return translation, scale, angle
